[FE] create: 추천 리스트 카드 정보 모달 파일 분리 생성
